+++ packages/connectkit/src/components/RozoPayModal/index.tsx
@@ -51,7 +51,6 @@ import WaitingDepositAddress from "../Pages/WaitingDepositAddress";
 import WaitingExternal from "../Pages/WaitingExternal";
 import WaitingWallet from "../Pages/WaitingWallet";
 import ConnectUsing from "./ConnectUsing";
-import ConnectWalletConnect from "../Pages/ConnectWalletConnect";
 
 export const RozoPayModal: React.FC<{
   mode: Mode;
@@ -135,8 +134,6 @@ export const RozoPayModal: React.FC<{
       context.setRoute(ROUTES.CONNECT, meta);
     } else if (context.route === ROUTES.CONNECTORS) {
       context.setRoute(ROUTES.SELECT_METHOD, meta);
-    } else if (context.route === ROUTES.CONNECT_WALLETCONNECT) {
-      context.setRoute(ROUTES.CONNECTORS, meta);
     } else if (context.route === ROUTES.SELECT_AMOUNT) {
       setSelectedTokenOption(undefined);
       context.setRoute(ROUTES.SELECT_TOKEN, meta);
@@ -280,7 +277,6 @@ export const RozoPayModal: React.FC<{
     [ROUTES.CONNECTORS]: <Connectors />,
     [ROUTES.MOBILECONNECTORS]: <MobileConnectors />,
     [ROUTES.CONNECT]: <ConnectUsing />,
-    [ROUTES.CONNECT_WALLETCONNECT]: <ConnectWalletConnect />,
     [ROUTES.SWITCHNETWORKS]: <SwitchNetworks />,
   };