+++ packages/connectkit/src/utils/index.ts
@@ -89,9 +89,6 @@ export const isSafeConnector = (connectorId?: string) => connectorId === "safe";
 export const isInjectedConnector = (connectorId?: string) =>
   connectorId === "injected";
 
-export const isWalletConnectConnector = (connectorId?: string) =>
-  connectorId === "walletConnect";
-
 export function parseError(e: any) {
   return {
     code: e?.error?.code || e?.code || -1,